adjtime: Fix wrong printf format when using CONFIG_SYSTEM_TIME64

+++ examples/adjtime/adjtime_main.c
@@ -148,7 +148,7 @@ int main(int argc, FAR char *argv[])
   parse_args(&delta, argc, argv);
 
   printf("Delta time is %ld seconds and %ld micro seconds.\n",
-         delta.tv_sec, delta.tv_usec);
+         (long)delta.tv_sec, delta.tv_usec);
 
   /* Call adjtime function. */
 
@@ -160,7 +160,7 @@ int main(int argc, FAR char *argv[])
   else
     {
       printf("Returned olddelta is %ld seconds and %ld micro seconds.\n",
-             olddelta.tv_sec, olddelta.tv_usec);
+             (long)olddelta.tv_sec, olddelta.tv_usec);
     }
 
   return ret;
